Separate app and admin into containers, finish reservation functionality
