+++ app/(main-page)/opengraph-image.alt.txt
@@ -1,0 +1,1 @@
+Dal logo
