+++ app/actions.ts
@@ -5,6 +5,7 @@ import { sql } from '@vercel/postgres'
 import { z } from 'zod'
 import { signIn } from '@/auth';
 import { AuthError } from 'next-auth';
+import { redirect } from 'next/navigation';
 
 export async function createReservation(prevState: any, formData: FormData) {
   const schema = z.object({
