--- app/admin/page.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: 'Login'
-}
-
-import LoginForm from "@/app/ui/layout/LoginForm"
-
-export default async function Admin() {
-  return(
-    <div className="flex w-screen h-screen bg-gray-900 items-center justify-center">
-      <LoginForm/>
-    </div>
-  )
-}
