--- app/opengraph-image.alt.txt
@@ -1,1 +1,0 @@
-Dal logo
