+++ auth.ts
@@ -4,6 +4,7 @@ import Credentials from 'next-auth/providers/credentials';
 import { z } from 'zod';
 import { sql } from '@vercel/postgres';
 import bcrypt from 'bcrypt';
+import { redirect } from 'next/navigation';
 
 type User = {
   id: string,
@@ -38,7 +39,9 @@ export const { auth, signIn, signOut } = NextAuth({
         if (!user) return null;
         const passwordsMatch = await bcrypt.compare(password, user.password);
 
-        if (passwordsMatch) return (user);
+        if (passwordsMatch) {
+          return (user)
+        };
       }
 
       console.log('Invalid credentials');
